# docs/swagger/swagger-sdce-1.json
{
  "openapi" : "3.0.1",
  "info" : {
    "contact" : {
      "email" : "onap-discuss@lists.onap.org",
      "name" : "ONAP",
      "url" : "https://onap.readthedocs.io"
    },
    "description" : "SDC API for onboarding (SDCE-1)",
    "license" : {
      "name" : "Apache 2.0",
      "url" : "http://www.apache.org/licenses/LICENSE-2.0"
    },
    "title" : "SPC API: SDCE-1",
    "version" : "1.0"
  },
  "servers" : [ {
    "description" : "SDCE-1 APIs",
    "url" : "/sdc"
  } ],
  "paths" : {
...
    "/v1.0/vendor-software-products/{vspId}/versions/{versionId}/deployment-flavors" : {
      "get" : {
        "description" : "List vendor software product Deployment Flavor",
        "operationId" : "list_8",
        "parameters" : [ {
          "description" : "Vendor software product Id",
          "in" : "path",
          "name" : "vspId",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        }, {
          "description" : "Version Id",
          "in" : "path",
          "name" : "versionId",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        }, {
          "in" : "header",
          "name" : "USER_ID",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "default" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "array",
                  "items" : {
                    "$ref" : "#/components/schemas/DeploymentFlavorListResponseDto"
                  }
                }
              }
            }
          }
        },
        "tags" : [ "SDCE-1 APIs", "Vendor Software Product deployment-flavors" ]
      },
      "post" : {
        "description" : "Create a vendor software product Deployment Flavor",
        "operationId" : "create_5",
        "parameters" : [ {
          "description" : "Vendor software product Id",
          "in" : "path",
          "name" : "vspId",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        }, {
          "description" : "Version Id",
          "in" : "path",
          "name" : "versionId",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        }, {
          "in" : "header",
          "name" : "USER_ID",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/DeploymentFlavorRequestDto"
              }
            }
          }
        },
        "responses" : {
          "default" : {
            "content" : {
              "application/json" : { }
            },
            "description" : "default response"
          }
        },
        "tags" : [ "SDCE-1 APIs", "Vendor Software Product deployment-flavors" ]
      }
    },
    "/v1.0/vendor-software-products/{vspId}/versions/{versionId}/deployment-flavors/schema" : {
      "get" : {
        "operationId" : "getSchema",
        "parameters" : [ {
          "description" : "Vendor software product Id",
          "in" : "path",
          "name" : "vspId",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        }, {
          "in" : "path",
          "name" : "versionId",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        }, {
          "in" : "header",
          "name" : "USER_ID",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "default" : {
            "content" : {
              "application/json" : { }
            },
            "description" : "default response"
          }
        },
        "tags" : [ "SDCE-1 APIs", "Vendor Software Product deployment-flavors" ]
      }
    },
    "/v1.0/vendor-software-products/{vspId}/versions/{versionId}/deployment-flavors/{deploymentFlavorId}" : {
      "delete" : {
        "description" : "Delete vendor software product Deployment Flavor",
        "operationId" : "delete_5",
        "parameters" : [ {
          "description" : "Vendor software product Id",
          "in" : "path",
          "name" : "vspId",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        }, {
          "description" : "Version Id",
          "in" : "path",
          "name" : "versionId",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        }, {
          "description" : "Vendor software product Deployment Flavor Id",
          "in" : "path",
          "name" : "deploymentFlavorId",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        }, {
          "in" : "header",
          "name" : "USER_ID",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "default" : {
            "content" : {
              "application/json" : { }
            },
            "description" : "default response"
          }
        },
        "tags" : [ "SDCE-1 APIs", "Vendor Software Product deployment-flavors" ]
      },
      "get" : {
        "description" : "Get vendor software product Deployment Flavor",
        "operationId" : "get_5",
        "parameters" : [ {
          "description" : "Vendor software product Id",
          "in" : "path",
          "name" : "vspId",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        }, {
          "description" : "Version Id",
          "in" : "path",
          "name" : "versionId",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        }, {
          "description" : "Vendor software product Deployment Flavor Id",
          "in" : "path",
          "name" : "deploymentFlavorId",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        }, {
          "in" : "header",
          "name" : "USER_ID",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "default" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/DeploymentFlavorDto"
                }
              }
            }
          }
        },
        "tags" : [ "SDCE-1 APIs", "Vendor Software Product deployment-flavors" ]
      },
      "put" : {
        "description" : "Update vendor software product Deployment Flavor",
        "operationId" : "update_4",
        "parameters" : [ {
          "description" : "Vendor software product Id",
          "in" : "path",
          "name" : "vspId",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        }, {
          "description" : "Version Id",
          "in" : "path",
          "name" : "versionId",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        }, {
          "description" : "Vendor software product Deployment Flavor Id",
          "in" : "path",
          "name" : "deploymentFlavorId",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        }, {
          "in" : "header",
          "name" : "USER_ID",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/DeploymentFlavorRequestDto"
              }
            }
          }
        },
        "responses" : {
          "default" : {
            "content" : {
              "application/json" : { }
            },
            "description" : "default response"
          }
        },
        "tags" : [ "SDCE-1 APIs", "Vendor Software Product deployment-flavors" ]
      }
    },
...
  },
  "components" : {
    "schemas" : {
...
      "ComponentComputeAssociation" : {
        "type" : "object",
        "properties" : {
          "componentId" : {
            "type" : "string"
          },
          "computeFlavorId" : {
            "type" : "string"
          }
        }
      },
...
      "DeploymentFlavorDto" : {
        "type" : "object",
        "properties" : {
          "componentComputeAssociations" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/ComponentComputeAssociation"
            }
          },
          "description" : {
            "type" : "string",
            "maxLength" : 300,
            "minLength" : 0
          },
          "featureGroupId" : {
            "type" : "string"
          },
          "id" : {
            "type" : "string"
          },
          "model" : {
            "type" : "string",
            "maxLength" : 30,
            "minLength" : 0
          }
        },
        "required" : [ "model" ]
      },
      "DeploymentFlavorListResponseDto" : {
        "type" : "object",
        "properties" : {
          "description" : {
            "type" : "string"
          },
          "id" : {
            "type" : "string"
          },
          "model" : {
            "type" : "string"
          }
        }
      },
      "DeploymentFlavorRequestDto" : {
        "type" : "object",
        "properties" : {
          "componentComputeAssociations" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/ComponentComputeAssociation"
            }
          },
          "description" : {
            "type" : "string",
            "maxLength" : 300,
            "minLength" : 0
          },
          "featureGroupId" : {
            "type" : "string"
          },
          "model" : {
            "type" : "string",
            "maxLength" : 30,
            "minLength" : 0
          }
        },
        "required" : [ "model" ]
      },
...
    }
  }
}